+++ src/fields/RadioSelect.vue
@@ -8,6 +8,15 @@
       :label="option.label"
       :group-name="name"
     />
+    <transition
+      name="radio-select__err-msg-transition"
+      @enter="setHeightCSSVar"
+      @before-leave="setHeightCSSVar"
+    >
+      <span v-if="errorMessage" class="radio-select__err-msg">
+        {{ errorMessage }}
+      </span>
+    </transition>
   </form>
 </template>
 
@@ -24,10 +33,12 @@ const props = withDefaults(
   defineProps<{
     name: string
     modelValue?: string
+    errorMessage?: string
     valueOptions: ValueOption[]
   }>(),
   {
     modelValue: '',
+    errorMessage: '',
   },
 )
 
@@ -36,6 +47,13 @@ const emit = defineEmits<{
 }>()
 
 const pickedValue = ref(props.modelValue)
+
+const setHeightCSSVar = (element: HTMLElement) => {
+  element.style.setProperty(
+    '--field-error-msg-height',
+    `${element.scrollHeight}px`,
+  )
+}
 
 watch(pickedValue, () => {
   emit('update:modelValue', pickedValue.value)
@@ -47,5 +65,30 @@ watch(pickedValue, () => {
   display: flex;
   flex-direction: column;
   gap: toRem(16);
+}
+
+.radio-select__err-msg {
+  @include field-error;
+
+  font-size: toRem(12);
+  padding-left: toRem(35);
+}
+
+.radio-select__err-msg-transition-enter-active {
+  animation: fade-down var(--field-transition-duration);
+}
+
+.radio-select__err-msg-transition-leave-active {
+  animation: fade-down var(--field-transition-duration) reverse;
+}
+
+@keyframes fade-down {
+  from {
+    height: 0;
+  }
+
+  to {
+    height: var(--field-error-msg-height);
+  }
 }
 </style>