+++ src/forms/purchase-book-form/default-payments/Erc20Template.vue
@@ -36,6 +36,7 @@
     </template>
   </template>
 
+  <!-- This component is teleported to parent component (Purchase Form) -->
   <teleport to="#purchase-book-form__preview">
     <book-preview :book="book" />
   </teleport>