+++ src/forms/purchase-book-form/default-payments/NativeTemplate.vue
@@ -15,6 +15,7 @@
   </template>
   <loader v-else />
 
+  <!-- This component is teleported to parent component (Purchase Form) -->
   <teleport to="#purchase-book-form__preview">
     <book-preview :book="book" />
   </teleport>